🎉 Readme Driven Development for ourOS
For a while now I've been thinking a major problem with the way we build
and deploy apps  is there's no real way for folks to operate programs on
systems they have  significant control over.

The move to the web solidified that. Languages like PHP held the line
for a while.

WordPress, for instance, is an incredibly powerful app development
platform that nearly anyone can run and develop for. And while PHP is
possibly one of the most accessible and widely used commercial
programming language, JavaScript is the _lingua franca_ of the modern
Web.

ourOS draws impact inspiration from WordPress. It's a human-centered,
pluggable multi-application hosting and distribution platform.

I believe that by building an _operating system_ that allows someone to
run multiple apps on a single host, we can decentralize the web and
break the hold that surveillance capitalism has on us.