Sprout Slipvector game
This is far from ready for play, but it's the start of a federated
command line / browser MUD that I've been noodling on for a while.

I figure starting super low fidelity will make it easier to design the
underlying game, and we can add fancier bits as time passes.